+++ src/gamePieces/Coordinate.java
@@ -1,4 +1,4 @@
-package GamePieces;
+package gamePieces;
 
 public class Coordinate {
